beta/fazai_genai_framework/requirements.txt
# GenAI Mini Framework - Dependências Python

# Core LLM and AI
google-generativeai>=0.8.0
openai>=1.51.0

# Vector Database
qdrant-client>=1.11.0

# Caching
gptcache>=0.1.44

# Embeddings and NLP
sentence-transformers>=3.1.0

# Utilities
python-dotenv>=1.0.1
